moving size_sat_seq funcs to own file in filters submodule
